Remove import script
