--- src/meshdb/utils/spreadsheet_import/logger.py
@@ -1,11 +1,0 @@
-import logging
-import os
-
-
-def configure():
-    FORMAT = "[%(levelname)s] %(message)s"
-
-    logging.basicConfig(
-        format=FORMAT,
-        level=logging.DEBUG if os.environ.get("LOG_DEBUG") else logging.INFO,
-    )
